+++ Client/Trivia/Source/Codec/S2C/Response/Packets/CreateRoomResponse.cs
@@ -4,6 +4,6 @@ using Trivia.Models.Raw;
 namespace Trivia.Codec.S2C.Response.Packets;
 
 public readonly record struct CreateRoomResponse(
-    int RoomId,
+    long RoomId,
     RoomData Data
 ) : IProtocolResponse;
